starter-kits/crewai/requirements.txt
crewai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
